Fix bugs

+++ Sources/AdaRender/Shaders/SpirvCompiler.swift
@@ -340,7 +340,7 @@ extension SpirvCompiler {
             let platform = SPVC_MSL_PLATFORM_IOS
 #endif
 
-            unsafe spvc_compiler_options_set_uint(options, SPVC_COMPILER_OPTION_MSL_PLATFORM, platform.rawValue)
+            unsafe spvc_compiler_options_set_uint(options, SPVC_COMPILER_OPTION_MSL_PLATFORM, UInt32(platform.rawValue))
             unsafe spvc_compiler_options_set_bool(options, SPVC_COMPILER_OPTION_MSL_ENABLE_DECORATION_BINDING, 1)
         }
